+++ deepspeed/launcher/launch.py
@@ -25,7 +25,7 @@ from typing import Dict
 from argparse import ArgumentParser, REMAINDER
 from ..constants import TORCH_DISTRIBUTED_DEFAULT_PORT
 from ..nebula.constants import DLTS_POD_ENV_PATH
-from ..utils import logger
+from ..utils import logger, get_numactl_cmd
 from ..elasticity import is_torch_elastic_compatible
 from .constants import ELASTIC_TRAINING_ID_DEFAULT
 
@@ -134,89 +134,6 @@ def terminate_process_tree(pid):
         p.kill()
 
 
-def parse_range(rng):
-    try:
-        value = int(rng)
-        return range(value, value + 1)
-    except ValueError:
-        # value is not a single number
-        parts = rng.split('-')
-        if len(parts) != 2:
-            raise ValueError("Bad range: '%s', range must be either a number or two number separated by dash" %
-                             (rng, ))
-        start = int(parts[0])
-        end = int(parts[1])
-        if start > end:
-            raise ValueError("Bad range: '%s', range end must larger than or equal to start" % (rng, ))
-        return range(start, end + 1)
-
-
-# parse comma and dash separated range list into list
-# i.e. "0,2-4,6" --> [0, 2, 3, 4, 6]
-# rules:
-# 1. Range list numser be comma sepeaated, each item are either a single number,
-#    or a range marked by two numbers (both number are included in the range)
-# 2. Sub ranges must be in ascend order and not overlap with each other
-# 3. No space in the range expression
-def parse_range_list(range_str):
-    number_list = []
-    last = -1
-    range_list = range_str.split(',')
-    for sub_range in range_list:
-        sub_number_list = parse_range(sub_range)
-        if sub_number_list[0] <= last:
-            raise ValueError(
-                "Bad range: '%s', sub ranges must not overlap with each other and should be in ascend order" %
-                (range_str, ))
-        last = sub_number_list[-1]
-        number_list.extend(sub_number_list)
-    return number_list
-
-
-# return a list of list for cores to numa mapping
-# [
-#     [ cores for numa 0 ]
-#     [ cores belong to numa 1 ]
-#     ...
-# ]
-def get_numa_cores():
-    ret = []
-    output = subprocess.check_output(['numactl', '--hardware']).decode("utf-8")
-    lines = output.split('\n')
-    for line in lines:
-        if line.startswith('available:'):
-            num_numas = int(line.split(' ')[1])
-            break
-    for numa in range(num_numas):
-        for line in lines:
-            if line.startswith(f'node {numa} cpus:'):
-                cores = line.split(' ')[3:]
-                ret.append([int(core) for core in cores])
-    return ret
-
-
-def check_for_numactl_pkg():
-    libs = dict(
-        dpkg=["-l", "numactl", "apt"],
-        pacman=["-Q", "numactl", "pacman"],
-        rpm=["-q", "numactl", "yum"],
-    )
-
-    found = False
-    for pkgmgr, data in libs.items():
-        flag, lib, tool = data
-        path = distutils.spawn.find_executable(pkgmgr)
-        if path is not None:
-            cmd = f"{pkgmgr} {flag} {lib}"
-            result = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-            if result.wait() == 0:
-                found = True
-            else:
-                print(f"please install the {lib} package with {tool}")
-            break
-    return found
-
-
 def main():
     args = parse_args()
     current_env = os.environ.copy()
@@ -313,39 +230,9 @@ def main():
             # spawn the processes
             cmd = []
             if args.bind_cores_to_rank:
-                check_for_numactl_pkg()
-                if 'KMP_AFFINITY' in os.environ.keys():
-                    raise ValueError("Environment variable KMP_AFFINITY conflicts with numactl "
-                                     "because it interfere with how many CPU cores numactl can set. "
-                                     "Unset KMP_AFFINITY before launching deepspeed.\n\n"
-                                     "\t$ unset KMP_AFFINITY\n"
-                                     "\t$ deepspeed <deepspeed command parameters>")
-                if args.bind_core_list != None:
-                    core_list = parse_range_list(args.bind_core_list)
-                    total_cores = len(core_list)
-                else:
-                    total_cores = psutil.cpu_count(logical=False)
-                    core_list = range(total_cores)
-                cores_per_rank = total_cores // num_local_procs
-                assert cores_per_rank >= 1, "At least one core needs to be assigned to each rank"
-                core_list_for_rank = core_list[cores_per_rank * local_rank:cores_per_rank * (local_rank + 1)]
+                cores_per_rank, numactl_cmd = get_numactl_cmd(args.bind_core_list, num_local_procs, local_rank)
                 current_env["OMP_NUM_THREADS"] = f"{cores_per_rank}"
-                cmd.append("numactl")
-
-                # check if all cores belong to same numa, if true, bind process to that numa domain with -m parameter
-                numa_cores = get_numa_cores()
-                num_numas = len(numa_cores)
-                for i in range(num_numas):
-                    if set(core_list_for_rank) <= set(numa_cores[i]):
-                        cmd.append("-m")
-                        cmd.append(f"{i}")
-                        break
-
-                cmd.append("-C")
-                core_list_str = f"{core_list_for_rank[0]}"
-                for core_id in core_list_for_rank[1:]:
-                    core_list_str = f"{core_list_str},{core_id}"
-                cmd.append(f"{core_list_str}")
+                cmd = cmd + numactl_cmd
             if not args.no_python:
                 cmd.append(sys.executable)
                 cmd.append("-u")
