+++ deepspeed/utils/__init__.py
@@ -14,3 +14,4 @@ from .tensor_fragment import tensor_fragment, get_full_hp_param, get_hp_fragment
 from .tensor_fragment import safe_get_full_fp32_param, safe_get_full_grad, safe_get_full_optimizer_state
 from .mixed_precision_linkage import link_hp_params
 from deepspeed.runtime.dataloader import RepeatingLoader
+from .numa import get_numactl_cmd
